Move WorkspaceSnapshot to core module.
