+++ model/src/main/java/com/chiralbehaviors/CoRE/workspace/WorkspaceSnapshot.java
@@ -18,7 +18,7 @@
  *  along with Ultrastructure.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.chiralbehaviors.CoRE.meta.workspace;
+package com.chiralbehaviors.CoRE.workspace;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -42,7 +42,6 @@ import org.slf4j.LoggerFactory;
 import com.chiralbehaviors.CoRE.Ruleform;
 import com.chiralbehaviors.CoRE.json.CoREModule;
 import com.chiralbehaviors.CoRE.product.Product;
-import com.chiralbehaviors.CoRE.workspace.WorkspaceAuthorization;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.core.JsonGenerationException;
